overhaul

+++ function/vibeCheckReallyHard.js
@@ -1,6 +1,9 @@
-module.exports =
-
-function (msg) {
-  client.destroy();
-  process.exit();
-}
+module.exports = async function (msg) {
+  if (msg && msg.author && msg.author.id !== davidUserID) {
+    return;
+  }
+  try {
+    if (global.client) client.destroy();
+  } catch (_) {}
+  process.exit(0);
+};
